requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.0.1
numpy==1.24.3
cryptography==41.0.7
dlib==19.24.2
cmake==3.27.7
